Implement Metal backend adapter with new renderer architecture

+++ src/renderers/generic/descriptions/BufferDescription.cs
@@ -4,7 +4,7 @@ namespace odl3d.Renderer;
 
 public sealed record BufferDescription
 {
-    public required nuint Size { get; init; }
+    public required int Size { get; init; }
     public required BufferUsage Usage { get; init; }
     public BufferAccess Access { get; init; } = BufferAccess.GpuOnly;
 }
